Make checks for undefined values more robust.
We had some issues with the nodelist breaking in case some values were
not `undefined`, but `null` instead.

The comparison `typeof value != 'undefined'` doesn't help here, so we
got errors like "f.autoupdater is null".

I replaced `typeof value != 'undefined'` with `value != null`, which
catches both `undefined` and `null` values, and seems to be a good way
to check for both those values, according to
https://stackoverflow.com/questions/2559318/how-to-check-for-an-undefined-or-null-variable-in-javascript